Moved classes to own files
